Added Search in Coupons, Blogs, Orders and DHL Token, Waybill, ShipmentInfo, Pathfinder added

+++ controller/variationController.js
@@ -10,6 +10,7 @@ const SubCategory = require("../models/subCategoryModel");
 const SpecialCategory = require("../models/specialCategory");
 const Size = require("../models/sizeModel");
 const Banner = require("../models/carouselModel");
+const Coupon = require("../models/couponModel");
 
 const config = {
   region: process.env.AWS_BUCKET_REGION,
@@ -43,6 +44,7 @@ const createCategory = asyncHandler(async (req, res) => {
 const updateCategory = asyncHandler(async (req, res) => {
   const { id, name, banner, image } = req.body;
   const ecomCategory = await Category.findById(id);
+ 
   if (ecomCategory) {
     ecomCategory.name = name;
     ecomCategory.banner = banner ? banner : ecomCategory.banner;
@@ -58,19 +60,32 @@ const updateCategory = asyncHandler(async (req, res) => {
 const deleteCategory = asyncHandler(async (req, res) => {
   const subid = req.query.id;
   const sub = await Category.findById(subid);
-
+  
   const f1 = sub.image;
+  
+  // if (f1) {
+  //   const fileName = f1.split("//")[1].split("/")[1];
 
-  if (f1) {
-    const fileName = f1.split("//")[1].split("/")[1];
+  //   const command = new DeleteObjectCommand({
+  //     Bucket: process.env.AWS_BUCKET,
+  //     Key: fileName,
+  //   });
+  //   const response = await s3.send(command);
+  // }
 
-    const command = new DeleteObjectCommand({
-      Bucket: process.env.AWS_BUCKET,
-      Key: fileName,
-    });
-    const response = await s3.send(command);
-  }
   // await Product.deleteMany({category: req.query.id})
+  const products = await Product.find({categoy: req.query.id})
+  if(products.length > 0) {
+    return res.status(400).send({status: false, message: "Delete all Products of this category first"})
+  }
+  const subcategories = await SubCategory.find({category: req.query.id})
+  if(subcategories.length > 0) {
+    return res.status(400).send({status: false, message: 'Delete SubCategories First'})
+  }
+  const specialCategories = await SpecialCategory.find({category: req.query.id}) 
+  if(specialCategories.length > 0) {
+    return res.status(400).send({status: false, message: 'Delete Special Categories First'})
+  }
   await Category.deleteOne({ _id: req.query.id });
   res.json("deleted");
 });
@@ -79,6 +94,17 @@ const getAllCategory = asyncHandler(async (req, res) => {
   const categories = await Category.find({});
   res.json(categories);
 });
+
+const getAllCategoryPaginationApplied = asyncHandler(async (req, res) => {
+  const pageNumber = Number(req.query.pageNumber) || 1
+  const pageSize = Number(req.query.pageSize) || 1  
+  const totalCategories = await Category.countDocuments({})
+  const pageCount = Math.ceil(totalCategories/pageSize)
+
+  const categories = await Category.find({}).skip((pageNumber - 1) * pageSize).limit(pageSize)
+
+  res.status(200).send({categories, pageCount})
+})
 
 const getSubCategoryByCategory = asyncHandler(async (req, res) => {
   const catId = req.query.catId;
@@ -130,16 +156,26 @@ const deleteSubCategory = asyncHandler(async (req, res) => {
 
   const f1 = sub.image;
 
-  if (f1) {
-    const fileName = f1.split("//")[1].split("/")[1];
+  // if (f1) {
+  //   const fileName = f1.split("//")[1].split("/")[1];
 
-    const command = new DeleteObjectCommand({
-      Bucket: process.env.AWS_BUCKET,
-      Key: fileName,
-    });
-    const response = await s3.send(command);
-  }
+  //   const command = new DeleteObjectCommand({
+  //     Bucket: process.env.AWS_BUCKET,
+  //     Key: fileName,
+  //   });
+  //   const response = await s3.send(command);
+  // }
   //   await Product.deleteMany({subcategory: req.query.id})
+
+  const products = await Product.find({subcategory: req.query.id})
+  if(products.length > 0) {
+    return res.status(400).send({status: false, message: "Delete all Products of this subcategory first"})
+  }
+
+  const specialCategories = await SpecialCategory.find({subcategory: req.query.id}) 
+  if(specialCategories.length > 0) {
+    return res.status(400).send({status: false, message: 'Delete Special Categories First'})
+  }
   await SubCategory.deleteOne({ _id: req.query.id });
   res.json("deleted");
 });
@@ -148,6 +184,16 @@ const getAllSubCategory = asyncHandler(async (req, res) => {
   const categories = await SubCategory.find({});
   res.json(categories);
 });
+
+const getAllSubCategoryPaginationApplied = asyncHandler(async(req, res) => {
+  const pageNumber = Number(req.query.pageNumber) || 1
+  const pageSize = Number(req.query.pageSize) || 1
+  const totalDocuments = await SubCategory.countDocuments({})
+
+  const pageCount = Math.ceil(totalDocuments/pageSize)
+  const subcategories = await SubCategory.find({}).skip((pageNumber - 1) * pageSize).limit(pageSize)
+  return res.status(200).send({status: true, categories: subcategories, pageCount})
+})
 
 //Special Category
 
@@ -175,6 +221,7 @@ const createSpecialCategory = asyncHandler(async (req, res) => {
 });
 const updateSpecialCategory = asyncHandler(async (req, res) => {
   const { id, name, subcategory, image } = req.body;
+
   const ecomCategory = await SpecialCategory.findById(id);
   const sub = await SubCategory.findById(subcategory);
   if (ecomCategory) {
@@ -192,20 +239,26 @@ const updateSpecialCategory = asyncHandler(async (req, res) => {
 });
 const deleteSpecialCategory = asyncHandler(async (req, res) => {
   const subid = req.query.id;
+  console.log(subid)
   const sub = await SpecialCategory.findById(subid);
-
+  console.log(sub)
   const f1 = sub.image;
+  console.log(f1)
+  // if (f1) {
+  //   const fileName = f1.split("//")[1].split("/")[1];
 
-  if (f1) {
-    const fileName = f1.split("//")[1].split("/")[1];
-
-    const command = new DeleteObjectCommand({
-      Bucket: process.env.AWS_BUCKET,
-      Key: fileName,
-    });
-    const response = await s3.send(command);
-  }
+  //   const command = new DeleteObjectCommand({
+  //     Bucket: process.env.AWS_BUCKET,
+  //     Key: fileName,
+  //   });
+  //   const response = await s3.send(command);
+  // }
   // await Product.deleteMany({category: req.query.id})
+  const products = await Product.find({specialcategory: req.query.id})
+  if(products.length > 0) {
+    return res.status(400).send({status: false, message: "Delete all Products of this special category first"})
+  }
+
   await SpecialCategory.deleteOne({ _id: req.query.id });
   res.json("deleted");
 });
@@ -216,6 +269,19 @@ const getAllSpecialCategory = asyncHandler(async (req, res) => {
 });
 
 //size
+
+const getAllSpecialCategoryPaginationApplied = asyncHandler(async(req, res) => {
+  const pageNumber = Number(req.query.pageNumber) || 1
+  const pageSize = Number(req.query.pageSize) || 1
+
+  const totalDocuments = await SpecialCategory.countDocuments({})
+  const pageCount = Math.ceil(totalDocuments/pageSize)
+
+  const specialcategories = await SpecialCategory.find({}).skip((pageNumber - 1) * pageSize).limit(pageSize)
+
+  return res.status(200).json({status: true, categories: specialcategories, pageCount})
+
+})
 
 const createSize = asyncHandler(async (req, res) => {
   const { id, name } = req.body;
@@ -243,6 +309,17 @@ const getAllSize = asyncHandler(async (req, res) => {
   res.json(categories);
 });
 
+const getAllSizePaginationApplied = asyncHandler(async (req, res) => {
+
+  const pageNumber = Number(req.query.pageNumber) || 1
+  const pageSize = Number(req.query.pageSize) || 1
+  const totalDocuments = await Size.countDocuments({})
+  
+  const pageCount = Math.ceil(totalDocuments/pageSize)
+  const size = await Size.find({}).skip((pageNumber  - 1) * pageSize).limit(pageSize)
+  
+  res.status(200).send({size, pageCount})
+})
 //homebanner
 
 const createBanner = asyncHandler(async (req, res) => {
@@ -272,6 +349,18 @@ const getBanner = asyncHandler(async (req, res) => {
     throw new Error("Error");
   }
 });
+
+const getBannerPaginationApplied = asyncHandler(async (req, res) => {
+  const pageNumber = Number(req.query.pageNumber) || 1
+  const pageSize = Number(req.query.pageSize) || 1
+  const totalDocuments = await Banner.countDocuments({})
+  
+  const pageCount = Math.ceil(totalDocuments/pageSize)
+  const banner = await Banner.find({}).skip((pageNumber  - 1) * pageSize).limit(pageSize)
+  res.status(200).send({banner, pageCount})
+})
+
+
 const deleteBanner = asyncHandler(async (req, res) => {
   const subid = req.query.id;
   const sub = await Banner.findById(subid);
@@ -286,6 +375,126 @@ const deleteBanner = asyncHandler(async (req, res) => {
   await Banner.deleteOne({ _id: req.query.id });
   res.json("deleted");
 });
+
+const searchCategory = asyncHandler(async (req, res) => {
+  const pageNumber = Number(req.query.pageNumber) || 1
+  const pageSize = Number(req.query.pageSize) || 1
+  const totalDocuments = await  Category.countDocuments({
+    $or: [
+      { name: { $regex: req.query.Query, $options: "i" } },
+      { _id: req.query.Query }  
+    ]
+  })
+
+  const pageCount = Math.ceil(totalDocuments/pageSize)
+
+  const categories = await Category.find({
+    $or: [
+      { name: { $regex: req.query.Query, $options: "i" } },
+      { _id: req.query.Query }  
+    ]
+  }).skip((pageNumber - 1) * pageSize).limit(pageSize)
+  
+  if (!categories || categories.length === 0) {
+    return res.status(404).json({ message: "No categories found" });
+  }
+  
+  res.status(200).json({categories, pageCount});
+})
+
+const searchSubCategory = asyncHandler(async (req, res) => {
+  const pageNumber = Number(req.query.pageNumber) || 1
+  const pageSize = Number(req.query.pageSize) || 1
+  const totalDocuments =  await SubCategory.countDocuments({
+    $or: [
+      { name: { $regex: req.query.Query, $options: "i" } },
+      { _id: req.query.Query },
+      { category: { $regex: req.query.Query, $options: "i" } }
+    ]
+  })
+
+  const pageCount = Math.ceil(totalDocuments/pageSize)
+
+  const subCategories = await SubCategory.find({
+    $or: [
+      { name: { $regex: req.query.Query, $options: "i" } },
+      { _id: req.query.Query },
+      { category: { $regex: req.query.Query, $options: "i" } }
+    ]
+  }).skip((pageNumber -1) * pageSize).limit(pageSize)
+  
+  if (!subCategories || subCategories.length === 0) {
+    return res.status(404).json({ message: "No subcategories found" });
+  }
+  
+  res.status(200).json({categories: subCategories, pageCount});
+  
+})
+
+const searchSpecialCategory = asyncHandler(async (req, res) => {
+  const pageNumber = Number(req.query.pageNumber) || 1
+  const pageSize = Number(req.query.pageSize) || 1
+  const totalDocuments = await SpecialCategory.find({
+    $or: [
+      { name: { $regex: req.query.Query, $options: "i" } },
+      { _id: req.query.Query },
+      { category: { $regex: req.query.Query, $options: "i" } },
+      { subcategory: { $regex: req.query.Query, $options: "i" } }
+    ]
+  })
+  const pageCount = Math.ceil(totalDocuments/pageSize)
+
+  const specialCategories = await SpecialCategory.find({
+    $or: [
+      { name: { $regex: req.query.Query, $options: "i" } },
+      { _id: req.query.Query },
+      { category: { $regex: req.query.Query, $options: "i" } },
+      { subcategory: { $regex: req.query.Query, $options: "i" } }
+    ]
+  }).skip((pageNumber - 1) * pageSize).limit(pageSize)
+  
+  if (!specialCategories || specialCategories.length === 0) {
+    return res.status(404).json({ message: "No special categories found" });
+  }
+  
+  res.status(200).json({categories: specialCategories, pageCount});
+  
+})
+
+
+
+const searchCoupons = asyncHandler(async (req, res) => {
+  
+  const query = req.query.Query || "";
+  const pageSize = 30;
+  const page = Number(req.query.pageNumber) || 1;
+  
+  const matchCriteria = {
+    $or: [
+      { name: { $regex: query, $options: "i" } }
+    ],
+  };
+
+  const count = await Coupon.countDocuments(matchCriteria);
+  const pageCount = Math.ceil(count / pageSize);
+ 
+  const coupons = await Coupon.find(matchCriteria)
+    .limit(pageSize)
+    .skip(pageSize * (page - 1))
+    .sort({ createdAt: -1 });
+   
+  if (!coupons || coupons.length === 0) {
+    return res.status(404).json({ message: "No coupons found" });
+  }
+ 
+  res.status(200).json({
+    coupons,
+    pageCount,
+  });
+});
+
+
+
 
 module.exports = {
   createBanner,
@@ -307,4 +516,13 @@ module.exports = {
   getAllSize,
   deleteSize,
   getSubCategoryByCategory,
+  getAllCategoryPaginationApplied,
+  getAllSpecialCategoryPaginationApplied,
+  getAllSubCategoryPaginationApplied,
+  searchCategory,
+  searchSubCategory,
+  searchSpecialCategory,
+  getAllSizePaginationApplied,
+  getBannerPaginationApplied,
+  searchCoupons
 };
